rename and move DocumentGroupDefinition
